refactor spinner page to components
